Refactor webview file handling and event management
- Consolidated file handling logic into a new module (fileHandlers.js) for improved organization and maintainability.
- Removed the obsolete fileHandler.js and updated imports across the codebase to utilize the new fileHandlers module.
- Streamlined event handling by introducing a dedicated messageHandlers module, enhancing clarity and reducing redundancy in script.js.
- Improved UI interaction by refactoring event listeners and state management functions, ensuring a smoother user experience.
- Enhanced modularity by separating concerns into distinct modules, facilitating easier future updates and maintenance.
